Implemented a new class numeric.InvertibleMatrix and improved test cases for ReplicaExchangeMC-derived classes.
The new InvertibleMatrix class is a matrix-like object which not only holds the
actual matrix, but also its inverse and is optimized on minimizing the number
of matrix inversions in the sampling algorithms. The sampling algorithms have been
optimized for this class.

The test cases for ReplicaExchangeMC-derived classes now sample from a multimodal
distribution which cannot be correctly sampled if there are serious bugs in the
algorithms which prevent sensible replica exchange.

The ThermostattedMDRENS test case for mass matrices different from unity was
trivialized such that the mass matrix employed now is a 1x1 unity matrix.
This makes the code treat it as if it was a non-trivial mass matrix, thus
the test is not completely useless, although there's no influence on the sampling.
This test will be improved by sampling from a specifically designed distribution.