Fix: fixed a minor bug in statistics.samplers.mc.multichain.AbstractStepRENS
The propagation parameters in the stepwise protocol were set such that the first propagation was performed with parameters of the unperturbed system instead of the parameters of the system after one perturbation step. This resulted in a
constant offset in the propagations of 1 "step" of the protocol, which makes little difference if there are many steps. Now it is fixed.